[Fix] adminView 구조 개선
