+++ src/main/java/com/cjoa/wms/util/DateRangeAction.java
@@ -1,0 +1,5 @@
+package com.cjoa.wms.util;
+@FunctionalInterface
+public interface DateRangeAction {
+    void execute(String startDate, String endDate);
+}
